+++ canadensis_core/src/lib.rs
@@ -34,7 +34,7 @@ pub use crate::error::{OutOfMemoryError, ServiceSubscribeError};
 const VALID_SUBJECT_IDS: RangeInclusive<u16> = 0..=8191;
 
 /// Subject ID, in range 0..=8191
-#[derive(Debug, Copy, Clone, Ord, PartialOrd, Eq, PartialEq, Hash)]
+#[derive(Copy, Clone, Ord, PartialOrd, Eq, PartialEq, Hash)]
 pub struct SubjectId(u16);
 
 impl SubjectId {
@@ -87,7 +87,7 @@ impl FromStr for SubjectId {
 const VALID_SERVICE_IDS: RangeInclusive<u16> = 0..=511;
 
 /// Service ID, in range 0..=511
-#[derive(Debug, Copy, Clone, Ord, PartialOrd, Eq, PartialEq, Hash)]
+#[derive(Copy, Clone, Ord, PartialOrd, Eq, PartialEq, Hash)]
 pub struct ServiceId(u16);
 
 impl ServiceId {
@@ -129,7 +129,7 @@ impl From<ServiceId> for usize {
 }
 
 /// A value that can represent a service ID (0..=511) or a subject ID (0..=8192)
-#[derive(Debug, Copy, Clone, Eq, PartialEq, Hash)]
+#[derive(Copy, Clone, Eq, PartialEq, Hash)]
 pub struct PortId(u16);
 
 impl From<SubjectId> for PortId {
@@ -234,10 +234,35 @@ impl TryFrom<u8> for Priority {
 }
 
 mod fmt_impl {
-    use super::SubjectId;
-    use core::fmt::{Display, Formatter, Result};
+    use super::{PortId, ServiceId, SubjectId};
+    use core::fmt::{Debug, Display, Formatter, Result};
 
+    impl Debug for SubjectId {
+        fn fmt(&self, f: &mut Formatter<'_>) -> Result {
+            Debug::fmt(&self.0, f)
+        }
+    }
     impl Display for SubjectId {
+        fn fmt(&self, f: &mut Formatter<'_>) -> Result {
+            Display::fmt(&self.0, f)
+        }
+    }
+    impl Debug for ServiceId {
+        fn fmt(&self, f: &mut Formatter<'_>) -> Result {
+            Debug::fmt(&self.0, f)
+        }
+    }
+    impl Display for ServiceId {
+        fn fmt(&self, f: &mut Formatter<'_>) -> Result {
+            Display::fmt(&self.0, f)
+        }
+    }
+    impl Debug for PortId {
+        fn fmt(&self, f: &mut Formatter<'_>) -> Result {
+            Debug::fmt(&self.0, f)
+        }
+    }
+    impl Display for PortId {
         fn fmt(&self, f: &mut Formatter<'_>) -> Result {
             Display::fmt(&self.0, f)
         }