+++ CHANGELOG.md
@@ -43,6 +43,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `canadensis_header`: Breaking change: Removed `header_crc()` public function and dependency on `crc-any`
 - Various crates: Breaking changes: Updated dependencies [#52](https://github.com/samcrow/canadensis/pull/52)
 - `canadensis_data_types`: Regenerated based on new public_regulated_data_types revision a229bb78e76c48a3082be163bc240b2c13ff2d89
+- `canadensis_core`: `SubjectId`, `ServiceId`, and `PortId` implement `Display` and a more compact form of `Debug`
+- `canadensis_can`: `CanNodeId`, and `CanTransferId` implement `Display` and a more compact form of `Debug`
+- `canadensis_header`: `NodeId16` and `SubjectId64` implement `Display` and a more compact form of `Debug`
 
 ### Fixed